Removed unused asset store adapters
